// src/api/response.type.ts
export interface GameTypeItem {
  id: GAME_TYPE.Enums
  game_type: string
  use_pc_image: boolean
  use_h5_image: boolean
  pc_image: string
  h5_image: string
  updated_at: number
  label?: string
  frontendKey?: string
}

export type GameTypeList = GameTypeItem[]

export interface GameTypeMap {
  [key: number]: GameTypeItem
}

export interface ProductItem {
  integration_id: number
  product_code: number
  product_name: string
  game_type: string
  square_image: string
  tab_image: string
  wide_image: string
  updated_at: number
}

export type ProductList = ProductItem[]

export interface GameItem {
  // game_type: string | number
  // product_name: string
  // updated_at: number
  game_code: string
  game_name: string
  game_type_id: number
  product_code: number
  favorite_count: number
  game_id: number
  hot: boolean
  newly: boolean
  sort: number
  is_favorite: boolean
  custom_image: string
  integration_id: number
  hot_count: number
  integration_created_at: number
  support_currency: string
}